* fixed (finally) the performance issue (related more to compiling tensorflow than to using it) * added "Expression" from the general tool box * added a new output that provides detailed info about state changes * added an Icon ;)

+++ SVDModel/SVDCore/core/externalseeds.cpp
@@ -124,7 +124,7 @@ int ExternalSeeds::setupFromStates(FileReader &rdr, Grid<int> &eseed)
 
         }
     }
-    return item_map.size();
+    return static_cast<int>(item_map.size());
 
 }
 
@@ -185,6 +185,6 @@ int ExternalSeeds::setupFromSpeciesShares(FileReader &rdr, Grid<int> &eseed)
 
         }
     }
-    return mExternalSeedTypes.size()-1; // without the "0"-type
+    return static_cast<int>(mExternalSeedTypes.size())-1; // without the "0"-type
 
 }
